Renamed existing service to zoteroAPI and added a second
The new service uses the `identity=1` flag to let zotero know that no
key needs to be created, and only account verification is wanted (for
mere login to bibbase).

TODO: merge zoteroAPI and zotero accounts post-authentication; it
should be possible to insert a hook there that checks the DB for an
existing zotero account for that (zotero) user id.